+++ SRC/NanoSpec/SpecifyCoordinateDlg.cpp
@@ -189,32 +189,32 @@ BOOL CSpecifyCoordinateDlg::CheckData()
 
 		switch ( m_StageConfig.OrgPos ) {
 		case 1: // 中心
-			coordLeftTop.lX = m_dSampleSizeX / 2 * -1;
-			coordLeftTop.lY = m_dSampleSizeY / 2;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX / 2 * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY / 2);
 			coordRightBottom.lX = m_dSampleSizeX / 2;
 			coordRightBottom.lY = m_dSampleSizeY / 2 * -1;
 			break;
 		case 2: // 左上
-			coordLeftTop.lX = 0;
-			coordLeftTop.lY = 0;
+			coordLeftTop.lX = static_cast<long>(0);
+			coordLeftTop.lY = static_cast<long>(0);
 			coordRightBottom.lX = m_dSampleSizeX;
 			coordRightBottom.lY = m_dSampleSizeY * -1;
 			break;
 		case 3: // 左下
-			coordLeftTop.lX = 0;
-			coordLeftTop.lY = m_dSampleSizeY;
+			coordLeftTop.lX = static_cast<long>(0);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
 			coordRightBottom.lX = m_dSampleSizeX;
 			coordRightBottom.lY = 0;
 			break;
 		case 4: // 右下
-			coordLeftTop.lX = m_dSampleSizeX * -1;
-			coordLeftTop.lY = m_dSampleSizeY;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
 			coordRightBottom.lX = 0;
 			coordRightBottom.lY = 0;
 			break;
 		case 5: // 右上
-			coordLeftTop.lX = m_dSampleSizeX * -1;
-			coordLeftTop.lY = 0;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
+			coordLeftTop.lY = static_cast<long>(0);
 			coordRightBottom.lX = 0;
 			coordRightBottom.lY = m_dSampleSizeY * -1;
 			break;
@@ -228,8 +228,8 @@ BOOL CSpecifyCoordinateDlg::CheckData()
 	///// 指定座標が、ステージ稼動範囲であるか
 	case SPECIFY_COORDINATE_STAGE:
 	default:
-		coordLeftTop.lX = m_StageConfig.StageMaxCoord.dLeftX;
-		coordLeftTop.lY = m_StageConfig.StageMaxCoord.dUpY;
+		coordLeftTop.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dLeftX);
+		coordLeftTop.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dUpY);
 		coordRightBottom.lX = m_StageConfig.StageMaxCoord.dRightX;
 		coordRightBottom.lY = m_StageConfig.StageMaxCoord.dDownY;
